Add ScaleDown.Actuator to AutoscalingContext
