+++ cluster-autoscaler/core/scaledown/deletiontracker/nodedeletiontracker.go
@@ -20,7 +20,7 @@ import (
 	"sync"
 	"time"
 
-	"k8s.io/autoscaler/cluster-autoscaler/processors/status"
+	"k8s.io/autoscaler/cluster-autoscaler/core/scaledown/status"
 	"k8s.io/autoscaler/cluster-autoscaler/utils/expiring"
 
 	apiv1 "k8s.io/api/core/v1"
